+++ src/algorithms/mod.rs
@@ -6,6 +6,7 @@
 
 pub mod connectivity;
 pub mod io;
+pub mod operators;
 pub mod paths;
 pub mod properties;
 pub mod traversal;